requirements.txt
aiohttp
azure-functions
azure-storage-blob
numpy
pandas==1.5.3